requirements.txt
absl-py==1.4.0
asttokens==2.2.1
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
chex==0.1.4
click==8.1.6
cloudpickle==2.2.1
colorama==0.4.6
commonmark==0.9.1
contextlib2==21.6.0
contourpy==1.1.0
cycler==0.11.0
Cython==0.29.36
d4rl @ git+https://github.com/ikostrikov/d4rl@99c475c7e187b370fa7621c6b2771d20421a4f18
decorator==5.1.1
distrax @ git+https://github.com/deepmind/distrax@776f60a2ec09cde6bc196e3afc9112ad28e9f590
dm-control==0.0.403778684
dm-env==1.6
dm-tree==0.1.8
docker-pycreds==0.4.0
etils==1.4.1
executing==1.2.0
fasteners==0.18
flatbuffers==1.12
flax==0.5.3
fonttools==4.42.0
future==0.18.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.32
glfw==2.6.2
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.57.0
gym==0.23.1
gym-notices==0.0.8
h5py==3.9.0
icecream==2.1.3
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.0.1
jax==0.3.15
jaxlib==0.3.15+cuda11.cudnn805
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
labmaze==1.0.6
libclang==16.0.6
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
metaworld @ git+https://github.com/rlworkgroup/metaworld.git@04be337a12305e393c0caf0cbf5ec7755c7c8feb
mjrl @ git+https://github.com/aravindr93/mjrl@3871d93763d3b49c4741e6daeaebbc605fe140dc
ml-collections==0.1.1
ml-dtypes==0.2.0
msgpack==1.0.5
mujoco==2.3.3
mujoco-py==2.1.2.14
numpy==1.22.4
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.1.3
packaging==23.1
patchelf==0.17.2.1
pathtools==0.1.2
Pillow==10.0.0
promise==2.3
protobuf==3.19.6
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybullet==3.2.5
pycparser==2.21
Pygments==2.16.1
PyOpenGL==3.1.7
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==11.2.0
rsa==4.9
scipy==1.11.1
sentry-sdk==1.29.2
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.33.0
tensorflow-probability==0.17.0
tensorstore==0.1.41
termcolor==2.3.0
toolz==0.12.0
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==1.26.16
wandb==0.13.1
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.16.2
